Changed footer section columns and added scroll to top button
html: Changed the ‘outro’ to 6 columns from 8, added 2 columns for the
‘scroll to top’ button {DESKTOP}. Changed ‘contact information’ to 10
columns on {MOBILE}. Added a href link for scroll to top button.

css: Added class for ‘scrollToTop’

js: Added function for when the ‘scrollToTop’ class is clicked on.
Animates the scroll back to top at a rate of 600 milliseconds.