Edited appearance of main controls
Added 4 SVG images for the controls of the site. Left and right for the
carousel control, up and down for the ‘hint’ and the ‘scroll to top’
button.

HTML:

The carousel edits have been executed using the ‘span class’ and
‘content’ method, the CSS tells the HTML file what ‘content’ it should
place in. (Line 101 + 105)

The up and down arrows are currently place in as <img src=“… tags but
could be placed in straight as SVG code. (?)

CSS:

The carousel controls in the css have been updated to now include
.arrow-left and .arrow-right under the .carouse-control class in order
for the new buttons to appear.

A large section of rules has been added under ARROW CONTROL
POSITIONING, to outline rules for the carousel arrows and the up and
down arrows. The up/down arrows have a rule which on large and x-large
screens makes them float right, but at medium, small and x-small they
float left, the default.

Background of .outro-content-wrap changed to dark grey colour.

JS:

Minimum window width for overlay has been changed to 750px.